fix: admin role, integer ratings, i18n, scan as collection entry
- set Kernic user role to Council on production database
- enforce integer ratings in backend and frontend
- add auth, scan, ratings i18n keys for de/en
- translate AuthBar, Scan, Layout, BeyDetail and PartDetail strings
- update app title to Beyblade X Database

+++ server/src/routes/catalog.ts
@@ -60,10 +60,10 @@ router.get('/parts/:category/:id', (req, res) => {
 });
 
 const ratingSchema = z.object({
-  attack: z.number().min(0).max(5),
-  defense: z.number().min(0).max(5),
-  stamina: z.number().min(0).max(5),
-  balance: z.number().min(0).max(5),
+  attack: z.number().int().min(0).max(5),
+  defense: z.number().int().min(0).max(5),
+  stamina: z.number().int().min(0).max(5),
+  balance: z.number().int().min(0).max(5),
 });
 
 router.post(
